test_axes is now part of test_series and test_dataframe
